+++ packages/mui-phone-input/src/default-metadata.d.ts
@@ -1,0 +1,6 @@
+declare module '#phone-default-metadata' {
+  import type { MetadataJson } from 'libphonenumber-js/core';
+
+  const metadata: MetadataJson;
+  export default metadata;
+}